Move App back to app module
